requirements.txt
absl-py
annotated-types
asgiref
astunparse
beautifulsoup4
cachetools
certifi
charset-normalizer
click
contourpy
cycler
Cython
dj-database-url
Django
django-heroku
django-jazzmin
djangorestframework
djongo
flatbuffers
fonttools
frozendict
gast
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
grpcio
grpcio-status
gunicorn
h5py
httplib2
idna
joblib
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
multitasking
namex
nltk
numpy
nvidia-nccl-cu12
opt_einsum
optree
packaging
pandas
patsy
peewee
pillow
platformdirs
pmdarima
proto-plus
protobuf
psycopg2
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pymongo
pyparsing
python-dateutil
python-dotenv
pytz
regex
requests
rich
rsa
scikit-learn
scipy
six
soupsieve
sqlparse
statsmodels
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tqdm
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
Werkzeug
whitenoise
wrapt
xgboost
yfinance
transformers
torch
